change folder structure of post component
