+++ src/Components/Post/Footer/CommentsModal.js
@@ -1,19 +1,19 @@
 import React, { Component } from "react";
-
-import getContentReplies from ".././Functions/getContentReplies";
-import sendComment from ".././Functions/sendComment";
+import Icon from "react-icons-kit";
+import { ic_message } from "react-icons-kit/md/ic_message";
+import Comments from "../../Comments/Comments";
+import getContentReplies from "Functions/getContentReplies";
+import sendComment from "Functions/sendComment";
 import uuidv4 from "uuid/v4";
-import delay from "../Functions/delay";
+import delay from "Functions/delay";
 
 import { hot } from "react-hot-loader";
-import colors from "../styles/colors";
+import colors from "styles/colors";
 import styled from "styled-components";
+import Modal from "react-modal";
 
-import store from "../store";
-import Comments from "../Components/Comments/Comments";
-const Container = styled.div`
-  overflow: hidden;
-`;
+import store from "../../../store";
+
 const Title = styled.div`
   box-sizing: border-box;
   text-align: center;
@@ -28,7 +28,6 @@ const Content = styled.div`
   box-sizing: border-box;
   height: 100%;
   width: 100%;
-  max-height: 30vh;
   overflow-y: scroll;
   padding-top: 20px;
   padding-bottom: 60px;
@@ -81,14 +80,16 @@ class CommentsModal extends Component {
     super(props);
 
     this.state = {
+      open: true,
       comments: [],
       comment: "",
       votedComments: [],
       votesToPush: [],
       votesToRemove: [],
+      isLoading: true,
       innerWidth: window.innerWidth
     };
-
+    this.handleOpen = this.handleOpen.bind(this);
     this.handleSendComment = this.handleSendComment.bind(this);
     this.handleInputChange = this.handleInputChange.bind(this);
     this.updateComments = this.updateComments.bind(this);
@@ -97,15 +98,8 @@ class CommentsModal extends Component {
   async componentWillMount() {
     window.addEventListener("resize", this.updateDimensions);
 
-    const apiCall = await getContentReplies(
-      this.props.postAuthor,
-      this.props.postPermlink
-    );
-
     this.setState({
-      comments: apiCall[0],
       open: true,
-      votedComments: await store.getState().steemProfileVotes.votes,
       isLoading: false
     });
   }
@@ -114,6 +108,29 @@ class CommentsModal extends Component {
       innerWidth: window.innerWidth
     });
   }
+  handleOpen = async () => {
+    const apiCall = await getContentReplies(
+      this.props.postAuthor,
+      this.props.postPermlink
+    );
+
+    await this.setState({
+      comments: apiCall[0],
+      open: true
+    });
+  };
+
+  handleClose = () => {
+    this.state.votesToPush.map(vote => {
+      this.updateVotingState(vote, true);
+      return void 0;
+    });
+
+    this.state.votesToRemove.map(vote => {
+      return this.updateVotingState(vote, false);
+    });
+    this.setState({ open: false });
+  };
 
   async updateComments() {
     await delay(3000);
@@ -158,37 +175,77 @@ class CommentsModal extends Component {
   }
 
   render() {
+    const modalStyle = {
+      content: {
+        left: "20px",
+        overflowY: "hidden",
+        overflowX: "hidden",
+        display: "flex",
+        bottom: "none",
+        flexDirection: "column",
+        padding: "0",
+        maxHeight: "60vh",
+        width:
+          this.state.innerWidth > 768
+            ? "40vw"
+            : this.state.innerWidth < 768 && this.state.innerWidth > 425
+              ? "60vw"
+              : "85vw"
+      },
+      overlay: {
+        backgroundColor: "transparent"
+      }
+    };
+
     return (
-      <Container>
-        <Title>
-          {this.props.likesNumber +
+      <span>
+        <Icon
+          icon={ic_message}
+          size={30}
+          style={{ cursor: "pointer" }}
+          onClick={this.handleOpen}
+        />
+        <Modal
+          title={
+            this.props.likesNumber +
             " Likes " +
             this.props.children +
-            " Comments"}
-        </Title>
+            " Comments"
+          }
+          modal={false}
+          isOpen={this.state.open}
+          onRequestClose={this.handleClose}
+          style={modalStyle}
+        >
+          <Title>
+            {this.props.likesNumber +
+              " Likes " +
+              this.props.children +
+              " Comments"}
+          </Title>
 
-        <Content>
-          <Comments
-            postAuthor={this.props.postAuthor}
-            postPermlink={this.props.postPermlink}
-          />
-        </Content>
-
-        <InputContainer>
-          <Input
-            placeholder="Send something"
-            name="comment"
-            type="text"
-            value={this.state.comment}
-            onChange={this.handleInputChange}
-          />
-          {this.state.comment.length > 1 ? (
-            <SendButton onClick={this.handleSendComment}>Reply</SendButton>
-          ) : (
-            <SendButton disabled>Reply</SendButton>
-          )}
-        </InputContainer>
-      </Container>
+          <Content>
+            <Comments
+              postAuthor={this.props.postAuthor}
+              postPermlink={this.props.postPermlink}
+            />
+          </Content>
+          <InputContainer>
+            <Input
+              placeholder="Send something"
+              name="comment"
+              type="text"
+              value={this.state.comment}
+              onChange={this.handleInputChange}
+            />
+            {this.state.comment.length > 1 ? (
+              <SendButton onClick={this.handleSendComment}>Reply</SendButton>
+            ) : (
+              <SendButton disabled>Reply</SendButton>
+            )}
+          </InputContainer>
+        </Modal>
+      </span>
     );
   }
 }
